dao: use long keys for text messages
Text message IDs, coming from the daemon, are 64 bits.
Store them as long in the database. This change is backward compatible.

+++ ring-android/app/src/main/java/cx/ring/history/DatabaseHelper.java
@@ -53,7 +53,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
     private static final int DATABASE_VERSION = 8;
 
     private Dao<HistoryCall, Integer> historyDao = null;
-    private Dao<HistoryText, Integer> historyTextDao = null;
+    private Dao<HistoryText, Long> historyTextDao = null;
 
     public DatabaseHelper(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -102,7 +102,7 @@ public class DatabaseHelper extends OrmLiteSqliteOpenHelper {
         return historyDao;
     }
 
-    public Dao<HistoryText, Integer> getTextHistoryDao() throws SQLException {
+    public Dao<HistoryText, Long> getTextHistoryDao() throws SQLException {
         if (historyTextDao == null) {
             historyTextDao = getDao(HistoryText.class);
         }
